% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchEvidenceExplorer}
\alias{launchEvidenceExplorer}
\title{Launch the Evidence Explorer Shiny app}
\usage{
launchEvidenceExplorer(
  cohorts = "class",
  dataFolder = "data",
  dataFile = "PreMerged.RData",
  connectionDetails = NULL,
  blind = TRUE,
  aboutText = NULL,
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)
}
\arguments{
\item{dataFolder}{A folder where the premerged file is stored. Use
the \code{\link{preMergeDiagnosticsFiles}} function to generate this file.}

\item{dataFile}{(Optional) The name of the .RData file with results. It is commonly known as the
Premerged file.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package, specifying how to connect to the server where
the study results results have been uploaded using the
\code{\link{uploadResults}} function.}

\item{blind}{Should effect-estimates be hidden?}

\item{aboutText}{Text (using HTML markup) that will be displayed in an About tab in the Shiny app.
If not provided, no About tab will be shown.}

\item{runOverNetwork}{(optional) Do you want the app to run over your network?}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}

\item{cohort}{A string denoting which results schema to explore.}
}
\description{
Launch the Evidence Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the evidence
}
\examples{
\dontrun{
# OHDSI shinydb legendt2dm read-only credentials
appConnectionDetails <- DatabaseConnector::createConnectionDetails(
  dbms = "postgresql",
  server = paste(keyring::key_get("legendt2dmServer"),
                 keyring::key_get("legendt2dmDatabase"),
                 sep = "/"),
  user = keyring::key_get("legendt2dmUser"),
  password = keyring::key_get("legendt2dmPassword"))

# Run from db server (data download can take a long time)
LegendT2dmEvidenceExplorer::launchEvidenceExplorer(cohorts = "class",
                                                   connectionDetails = appConnectionDetails,
                                                   blind = TRUE)
}

}
